+++ lib/src/test/kotlin/app/cash/nostrino/model/EventTest.kt
@@ -16,22 +16,12 @@
 
 package app.cash.nostrino.model
 
-import app.cash.nostrino.ArbPrimitive.arbByteString32
-import app.cash.nostrino.ArbPrimitive.arbByteString64
-import app.cash.nostrino.ArbPrimitive.arbInstantSeconds
 import app.cash.nostrino.crypto.SecKeyGenerator
 import app.cash.nostrino.message.NostrMessageAdapter.Companion.moshi
-import app.cash.nostrino.model.ArbEventContent.arbEncryptedDm
-import app.cash.nostrino.model.ArbEventContent.arbReaction
+import app.cash.nostrino.model.ArbEvent.arbEventWithContent
 import app.cash.nostrino.model.ArbEventContent.arbTextNote
-import app.cash.nostrino.model.ArbEventContent.arbUserMetaData
-import app.cash.nostrino.model.ArbEventContent.arbZapRequest
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
-import io.kotest.property.Arb
-import io.kotest.property.arbitrary.bind
-import io.kotest.property.arbitrary.choice
-import io.kotest.property.arbitrary.map
 import io.kotest.property.arbitrary.next
 import io.kotest.property.checkAll
 
@@ -77,36 +67,4 @@ class EventTest : StringSpec({
 
     event?.validSignature shouldBe true
   }
-}) {
-  companion object {
-    private val arbEventContent = Arb.choice(
-      arbTextNote,
-      arbEncryptedDm,
-      arbUserMetaData,
-      arbReaction,
-      arbZapRequest
-    )
-
-    val arbEventWithContent: Arb<Pair<Event, EventContent>> by lazy {
-      Arb.bind(
-        arbByteString32,
-        arbByteString32,
-        arbInstantSeconds,
-        arbEventContent,
-        arbByteString64
-      ) { id, pubKey, createdAt, content, sig ->
-        val event = Event(
-          id = id,
-          pubKey = pubKey,
-          createdAt = createdAt,
-          kind = content.kind,
-          tags = content.tags.map { it.toJsonList() },
-          content = content.toJsonString(),
-          sig = sig
-        )
-        event to content
-      }
-    }
-    val arbEvent: Arb<Event> by lazy { arbEventWithContent.map { it.first } }
-  }
-}
+})
