+++ lib/src/test/kotlin/app/cash/nostrino/model/ZapRequestTest.kt
@@ -1,6 +1,6 @@
 package app.cash.nostrino.model
 
-import app.cash.nostrino.crypto.SecKeyTest.Companion.arbSecKey
+import app.cash.nostrino.crypto.ArbKeys.arbSecKey
 import app.cash.nostrino.message.NostrMessageAdapter
 import app.cash.nostrino.model.ArbEventContent.arbZapRequest
 import io.kotest.core.spec.style.StringSpec
